#7: Added a Console Application
In this commit, we added a simple console application that echoes the input. The ChatConsole class implements the Runnable interface because we want to run this app in a separate Thread. In the run method, we have an infinite while because we want to run this app forever. In this loop, the program is waiting for user input, and it prints the user input to the console whenever the user inserts a line in the console as an input.