#7: Added the send message functionality to the Client Console
To add the send message functionality to the console version of the client, we use the ChatAccess class, which is responsible for sending a message to the server. We changed the logic of the part that is responsible for the reading line from the client console. We removed the echo code, and we call the send method to send the string that the client just wrote to the server.